requirements.txt
asgiref==3.2.5
Django==3.0.4
djangorestframework==3.11.0
Pillow==7.0.0
pytz
six
sqlparse==0.3.1
